requirements.railway.txt
boto3>=1.38.0
botocore>=1.38.0
python-dotenv==1.0.1
anthropic==0.34.0
pydantic==2.8.0
requests==2.32.0